requirements.txt
fastapi
uvicorn
requests
sqlalchemy
python-dotenv
pytrends
pydantic
